requirements.txt
Django==5.2.9
django-browser-reload==1.21.0
psycopg2==2.9.11
django-yamlconf==1.5.0
gdal-installer==1.0.2
numpy==1.26.4
django-htmx==1.8.0
python-dateutil==2.9.0
django-crispy-forms==2.5
crispy-tailwind==1.0.3